ProductPage37

+++ server/routes/multer.js
@@ -67,6 +67,11 @@ router.post("/upload/:tpye/:name", (req, res) => {
   const type = req.params.tpye;
   const name = req.params.name;
 
+  if (req?.file?.length === 0) {
+    // 파일 선택되지 않은 경우
+    return res.status(400).json({ success: false, message: "파일을 선택해주세요." });
+  }
+
   try {
     fs.readdirSync(`../client/public/img/${type}/${name}`); // 폴더 확인
   } catch (err) {
@@ -97,17 +102,21 @@ router.post("/upload/:tpye/:name", (req, res) => {
 
   upload(req, res, (err) => {
     if (err) {
-      return res.json({ success: false, err });
+      return res.status(400).json({ success: false, err });
     }
-    if (res.req.file.path.length === 0) {
+    if (res.req?.file?.path?.length === 0) {
       return;
+    }
+    const regex = new RegExp(`undefined`, "g");
+    if (regex.test(res.req?.file?.filename)) {
+      return res.status(400).json({ success: false, message: "파일을 선택해주세요." });
     }
     return res.json({
       success: true,
-      image: res.req.file.path,
+      image: res.req?.file?.path,
       imagePath: `/img/${type}/${name}`,
-      imagePathName: `/img/${type}/${name}/${res.req.file.filename}`,
-      fileName: res.req.file.filename,
+      imagePathName: `/img/${type}/${name}/${res.req?.file?.filename}`,
+      fileName: res.req?.file?.filename,
     });
   });
   router.use(`../client/public/img/${type}/${name}`, express.static(path.join(__dirname, `../client/public/img/${type}/${name}`)));
@@ -119,6 +128,9 @@ router.post("/delete/route", async (req, res) => {
     res.status(403).send(`null`);
   } else {
     const directory = fs.existsSync(`../client/public/${route}`);
+    if (!directory) {
+      return res.status(404).send("Not Found");
+    }
     if (directory) {
       try {
         fs.rmdirSync(`../client/public/${route}`, { recursive: true });
@@ -148,6 +160,13 @@ router.post("/delete/fill", async (req, res) => {
       console.log(error);
     }
   }
+});
+
+router.put("/route/put", async (req, res) => {
+  fs.rename(`../client/public${req.body.route}`, `../client/public${req.body.newRoute}`, (err) => {
+    if (err) throw err;
+    res.status(201).send("변경 완료");
+  });
 });
 
 module.exports = router;
